# specs/001-gs-control-center/contracts/api.yaml
openapi: 3.1.0
info:
  title: GS Control Center API
  description: Centralized Advertising Block Management System for UCS Guest Screen
  version: 1.0.0
servers:
  - url: /api/v1
    description: Central Server API (Host 10.0.0.111)
paths:
...
  /regions:
    get:
      summary: List all regions
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Region'
    post:
      summary: Create new region
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, code]
              properties:
                name: { type: string }
                code: { type: string }
                default_full_screen_block_id: { type: string, format: uuid, nullable: true }
                default_mode32_block_id: { type: string, format: uuid, nullable: true }
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Region'
  /branches:
    get:
      summary: List branches (filterable by region)
      parameters:
        - name: region_id
          in: query
          schema: { type: string, format: uuid }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Branch'
    post:
      summary: Create new branch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [region_id, name, code]
              properties:
                region_id: { type: string, format: uuid }
                name: { type: string }
                code: { type: string }
                override_full_screen_block_id: { type: string, format: uuid, nullable: true }
                override_mode32_block_id: { type: string, format: uuid, nullable: true }
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Branch'
  /cashiers:
    get:
      summary: List cashiers
      parameters:
        - name: branch_id
          in: query
          schema: { type: string, format: uuid }
        - name: status
          in: query
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cashier'
    post:
      summary: Register new cashier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [branch_id, name, ip_address]
              properties:
                branch_id: { type: string, format: uuid }
                name: { type: string }
                ip_address: { type: string, format: ipv4 }
                ssh_port: { type: integer, default: 22 }
                override_full_screen_block_id: { type: string, format: uuid, nullable: true }
                override_mode32_block_id: { type: string, format: uuid, nullable: true }
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cashier'
  /cashiers/{id}/test-connection:
    post:
      summary: Test SSH connectivity to cashier
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  reachable: { type: boolean }
                  response_time_ms: { type: integer }
                  gs_installed: { type: boolean }
                  free_space_mb: { type: integer }
                  error_message: { type: string, nullable: true }
...
components:
  schemas:
    Region:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        code: { type: string }
        default_full_screen_block_id: { type: string, format: uuid, nullable: true }
        default_mode32_block_id: { type: string, format: uuid, nullable: true }
        created_at: { type: string, format: date-time }
    Branch:
      type: object
      properties:
        id: { type: string, format: uuid }
        region_id: { type: string, format: uuid }
        name: { type: string }
        code: { type: string }
        override_full_screen_block_id: { type: string, format: uuid, nullable: true }
        override_mode32_block_id: { type: string, format: uuid, nullable: true }
    Cashier:
      type: object
      properties:
        id: { type: string, format: uuid }
        branch_id: { type: string, format: uuid }
        name: { type: string }
        ip_address: { type: string }
        ssh_port: { type: integer }
        enabled: { type: boolean }
        has_ssh_password: { type: boolean, description: Indicates whether an encrypted SSH password is configured }
        last_sync_status: { type: string }
        last_seen_at: { type: string, format: date-time, nullable: true }
...